+++ index.html
@@ -229,7 +229,9 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.20/lodash.min.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
+  <script src="js/data.js" type="module"></script>
+  <script src="js/utils/get-random-positive-integer.js" type="module"></script>
+  <script src="js/utils/check-string-length.js" type="module"></script>
 </body>
 </html>
